requirements.txt
requests>=2.28.0
Pillow>=9.0.0
pyzbar>=0.1.9
opencv-python>=4.5.0
